create Authentication

+++ app/Http/Controllers/IdeaController.php
@@ -29,13 +29,19 @@ class IdeaController extends Controller
         // );
 
         //$idea = Idea::create(request()->all()); //сокращенная форма, требующая $quarded или $fillable поля в модели 
-        $idea = Idea::create($validated);
+        $validated['user_id'] = auth()->id();
+
+        Idea::create($validated);
 
         return redirect()->route('dashboard')->with('success', 'Idea created successfully!');
     }
 
     public function destroy(Idea $idea)
     {
+        if (auth()->id() !== $idea->user_id) {
+            abort(404);
+        }
+
         $idea->delete();
 
         return redirect()->route('dashboard')->with('success', 'Idea deleted successfully!');
@@ -43,12 +49,20 @@ class IdeaController extends Controller
 
     public function edit(Idea $idea)
     {
+        if (auth()->id() !== $idea->user_id) {
+            abort(404);
+        }
+
         $editing = true;
         return view("ideas.show", compact("idea", "editing"));
     }
 
     public function update(Idea $idea)
     {
+        if (auth()->id() !== $idea->user_id) {
+            abort(404);
+        }
+
         $validated = request()->validate([
             'content' => 'required|min:5|max:240',
         ]);
